requirements_app.txt
wheel
fastapi==0.95.2
uvicorn==0.22.0
tensorflow==2.16.2
numpy==1.26.0
#rasterio==1.2.10
pydantic==1.10.8
pillow==9.5.0
